fix: rethink when to attach to project and not solution
There is a couple of cases where I want to attach to the csproj and not
sln. One example is if I for example open up neovim and I don't have any
sln files in cwd or in any subdirectories.

In this case I want to use csproj instead of sln file.

This is more in line with how other editors work with `workspaces`, as
it takes into account where you open the editor.

However, I don't want it to use the .csproj if I open up neovim in a
folder, and I then for some reason open up a .cs file in a totally
different project and solution.