feat: add silent option
Allow to opt out of the messages printed by this plugin that are used
purely for notifying the user of initialization.

There are other notifications potentially printed by this plugin, but I
strictly believe not _all_ notifications should be affected by the
silent option. The other ones are mostly used if something goes wrong
during an explicit action by the user, or notify the user about an
action they need to do