feat(health): Razor checks (#274)
* feat(health): check for Razor extension presence

Add a health check to verify the presence of the Razor extension in the
Mason roslyn package. Warn the user if the extension is not found,
indicating that Razor support will be limited and suggesting installation
via Mason.

* feat(issue-template): update minimal config instructions

Rename minimal config file to `repro.lua` and update instructions to use
`nvim -u repro.lua` for reproduction. Switch to using `lazy.minit` for
plugin setup and add Mason registry details. Clarify steps for Roslyn
language server installation and filetype setup.

* feat(health): add checks for html-lsp and server

Added health checks for the presence of vscode-html-language-server and
html-lsp client configuration. Warns if either is missing, guiding users
to improve Razor HTML support by installing or configuring the necessary
components.

* refactor(github): remove lazyloading and config fn